fix msys-support completely, fully functional
