+++ settings.go
@@ -50,17 +50,9 @@ func extensionsDir() (string, error) {
 }
 
 func windowsVarFix(input string) string {
-	var remoteSettingsDir = "~/.local/share/code-server/User/"
-	var remoteExtensionsDir = "~/.local/share/code-server/extensions/"
 	if runtime.GOOS == "windows" {
-		remoteSettingsDir = ".local/share/code-server/User/"
-		remoteExtensionsDir = ".local/share/code-server/extensions/"
+		return input
 	}
 
-	if input == "remoteSettingsDir" {
-		return remoteSettingsDir
-	} else if input == "remoteExtensionsDir" {
-		return remoteExtensionsDir
-	}
-	return ""
+	return "~/" + input
 }
